update from origin

+++ kbe/src/server/tools/bots/bots.cpp
@@ -379,7 +379,7 @@ PyObject* Bots::__py_addBots(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------	
 PyObject* Bots::__py_setScriptLogType(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	int argCount = (int)PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::scriptLogType(): args is error!");
